move with-registration to components
