+++ packages/template-retail-react-app/app/components/product-view/index.jsx
@@ -20,7 +20,7 @@ import Swatch from '../swatch-group/swatch'
 import ImageGallery from '../image-gallery'
 import Breadcrumb from '../breadcrumb'
 import Link from '../link'
-import withRegistration from '../../hoc/with-registration'
+import withRegistration from '../with-registration'
 import {useCurrency} from '../../hooks'
 import {Skeleton as ImageGallerySkeleton} from '../image-gallery'
 import {HideOnDesktop, HideOnMobile} from '../responsive'
