+++ packages/template-retail-react-app/app/components/with-registration/index.jsx
